+++ polardb-20170801/src/Models/RegisterKnowledgeBaseFileResponseBody.php
@@ -1,0 +1,118 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Polardb\V20170801\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class RegisterKnowledgeBaseFileResponseBody extends Model
+{
+    /**
+     * @var string
+     */
+    public $fileId;
+
+    /**
+     * @var string
+     */
+    public $fileName;
+
+    /**
+     * @var string
+     */
+    public $filePath;
+
+    /**
+     * @var bool
+     */
+    public $idempotent;
+
+    /**
+     * @var string
+     */
+    public $requestId;
+
+    /**
+     * @var string
+     */
+    public $status;
+    protected $_name = [
+        'fileId' => 'FileId',
+        'fileName' => 'FileName',
+        'filePath' => 'FilePath',
+        'idempotent' => 'Idempotent',
+        'requestId' => 'RequestId',
+        'status' => 'Status',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->fileId) {
+            $res['FileId'] = $this->fileId;
+        }
+
+        if (null !== $this->fileName) {
+            $res['FileName'] = $this->fileName;
+        }
+
+        if (null !== $this->filePath) {
+            $res['FilePath'] = $this->filePath;
+        }
+
+        if (null !== $this->idempotent) {
+            $res['Idempotent'] = $this->idempotent;
+        }
+
+        if (null !== $this->requestId) {
+            $res['RequestId'] = $this->requestId;
+        }
+
+        if (null !== $this->status) {
+            $res['Status'] = $this->status;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['FileId'])) {
+            $model->fileId = $map['FileId'];
+        }
+
+        if (isset($map['FileName'])) {
+            $model->fileName = $map['FileName'];
+        }
+
+        if (isset($map['FilePath'])) {
+            $model->filePath = $map['FilePath'];
+        }
+
+        if (isset($map['Idempotent'])) {
+            $model->idempotent = $map['Idempotent'];
+        }
+
+        if (isset($map['RequestId'])) {
+            $model->requestId = $map['RequestId'];
+        }
+
+        if (isset($map['Status'])) {
+            $model->status = $map['Status'];
+        }
+
+        return $model;
+    }
+}