+++ polardb-20170801/src/Models/RetrievalKnowledgeBaseResponseBody/results.php
@@ -5,9 +5,20 @@
 namespace AlibabaCloud\SDK\Polardb\V20170801\Models\RetrievalKnowledgeBaseResponseBody;
 
 use AlibabaCloud\Dara\Model;
+use AlibabaCloud\SDK\Polardb\V20170801\Models\RetrievalKnowledgeBaseResponseBody\results\imageResources;
 
 class results extends Model
 {
+    /**
+     * @var string[]
+     */
+    public $captions;
+
+    /**
+     * @var string[]
+     */
+    public $docItems;
+
     /**
      * @var string
      */
@@ -22,6 +33,11 @@ class results extends Model
      * @var string[]
      */
     public $headings;
+
+    /**
+     * @var imageResources[]
+     */
+    public $imageResources;
 
     /**
      * @var string
@@ -48,9 +64,12 @@ class results extends Model
      */
     public $similarityScore;
     protected $_name = [
+        'captions' => 'Captions',
+        'docItems' => 'DocItems',
         'fileId' => 'FileId',
         'fileName' => 'FileName',
         'headings' => 'Headings',
+        'imageResources' => 'ImageResources',
         'metadata' => 'Metadata',
         'pageNumbers' => 'PageNumbers',
         'shardContent' => 'ShardContent',
@@ -60,8 +79,17 @@ class results extends Model
 
     public function validate()
     {
+        if (\is_array($this->captions)) {
+            Model::validateArray($this->captions);
+        }
+        if (\is_array($this->docItems)) {
+            Model::validateArray($this->docItems);
+        }
         if (\is_array($this->headings)) {
             Model::validateArray($this->headings);
+        }
+        if (\is_array($this->imageResources)) {
+            Model::validateArray($this->imageResources);
         }
         if (\is_array($this->pageNumbers)) {
             Model::validateArray($this->pageNumbers);
@@ -72,6 +100,28 @@ class results extends Model
     public function toArray($noStream = false)
     {
         $res = [];
+        if (null !== $this->captions) {
+            if (\is_array($this->captions)) {
+                $res['Captions'] = [];
+                $n1 = 0;
+                foreach ($this->captions as $item1) {
+                    $res['Captions'][$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (null !== $this->docItems) {
+            if (\is_array($this->docItems)) {
+                $res['DocItems'] = [];
+                $n1 = 0;
+                foreach ($this->docItems as $item1) {
+                    $res['DocItems'][$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
         if (null !== $this->fileId) {
             $res['FileId'] = $this->fileId;
         }
@@ -86,6 +136,17 @@ class results extends Model
                 $n1 = 0;
                 foreach ($this->headings as $item1) {
                     $res['Headings'][$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (null !== $this->imageResources) {
+            if (\is_array($this->imageResources)) {
+                $res['ImageResources'] = [];
+                $n1 = 0;
+                foreach ($this->imageResources as $item1) {
+                    $res['ImageResources'][$n1] = null !== $item1 ? $item1->toArray($noStream) : $item1;
                     ++$n1;
                 }
             }
@@ -129,6 +190,28 @@ class results extends Model
     public static function fromMap($map = [])
     {
         $model = new self();
+        if (isset($map['Captions'])) {
+            if (!empty($map['Captions'])) {
+                $model->captions = [];
+                $n1 = 0;
+                foreach ($map['Captions'] as $item1) {
+                    $model->captions[$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (isset($map['DocItems'])) {
+            if (!empty($map['DocItems'])) {
+                $model->docItems = [];
+                $n1 = 0;
+                foreach ($map['DocItems'] as $item1) {
+                    $model->docItems[$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
         if (isset($map['FileId'])) {
             $model->fileId = $map['FileId'];
         }
@@ -143,6 +226,17 @@ class results extends Model
                 $n1 = 0;
                 foreach ($map['Headings'] as $item1) {
                     $model->headings[$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (isset($map['ImageResources'])) {
+            if (!empty($map['ImageResources'])) {
+                $model->imageResources = [];
+                $n1 = 0;
+                foreach ($map['ImageResources'] as $item1) {
+                    $model->imageResources[$n1] = imageResources::fromMap($item1);
                     ++$n1;
                 }
             }